maybe fix 250% engine load

+++ src/data.rs
@@ -3,5 +3,5 @@ pub struct GaugeData {
     pub oil_temp_est: f32,
     pub coolant_temp: f32,
     pub voltage: f32,
-    pub engine_load: u8,
+    pub engine_load: f32,
 }
